# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Management API
  version: 2.0.0
  description: |
    ระบบบริหารคลังสินค้า เน้นความถูกต้องของยอดสต็อกและการบันทึกประวัติทุกครั้งที่มีการปรับ

    ทุก request ที่มี body ต้องส่ง header `Content-Type: application/json`
servers:
  - url: http://localhost:3000
    description: เครื่องพัฒนา
...
paths:
...
  /api/stock/adjust:
    patch:
      tags: [Stock]
      summary: ปรับเพิ่มหรือลดสต็อก
      description: |
        ค่าบวกคือรับเข้า ค่าลบคือตัดออก ห้ามเป็น 0

        ทำงานใน transaction เดียว ล็อกแถวสินค้าด้วย `SELECT ... FOR UPDATE`
        ถ้าผลลัพธ์ทำให้สต็อกติดลบ จะตอบ 409 และไม่มีการเปลี่ยนแปลงใด ๆ
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StockAdjustRequest' }
            examples:
              receive:
                summary: รับของเข้าคลัง
                value: { product_id: 2, quantity: 10, reason: รับของจากซัพพลายเออร์, created_by: somchai }
              issue:
                summary: ตัดออกจากคลัง
                value: { product_id: 2, quantity: -5, reason: เบิกใช้ภายใน }
      responses:
        '200':
          description: ปรับสำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      product: { type: object }
                      adjustment:
                        type: object
                        properties:
                          balance_before: { type: integer, example: 3 }
                          change: { type: integer, example: 5 }
                          balance_after: { type: integer, example: 8 }
                      transaction: { $ref: '#/components/schemas/StockTransaction' }
                  message: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: สต็อกไม่พอ หรือสินค้าถูกปิดการใช้งาน
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: INSUFFICIENT_STOCK
                  message: 'สต็อกไม่พอ: เมาส์ไร้สาย Logitech M331 คงเหลือ 3 ตัว แต่ขอตัด 5 ตัว'
                  details: { product_id: 2, sku: IT-MS-002, stock_quantity: 3, requested: -5, shortfall: 2 }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/stock/bulk-adjust:
    patch:
      tags: [Stock]
      summary: ปรับสต็อกหลายรายการในครั้งเดียว (ทั้งชุดสำเร็จหรือยกเลิกทั้งหมด)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { $ref: '#/components/schemas/StockAdjustRequest' }
      responses:
        '200': { description: สำเร็จทุกรายการ }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /api/stock/transactions:
    get:
      tags: [Stock]
      summary: ประวัติการเคลื่อนไหวทั้งระบบ
      parameters:
        - { name: product_id, in: query, schema: { type: integer } }
        - { name: type, in: query, schema: { type: string, enum: [IN, OUT] } }
        - { name: from, in: query, schema: { type: string, format: date-time }, example: '2026-09-01' }
        - { name: to, in: query, schema: { type: string, format: date-time }, example: '2026-09-30' }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/StockTransaction' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
components:
...
  responses:
    NotFound:
      description: ไม่พบข้อมูล
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: PRODUCT_NOT_FOUND, message: ไม่พบสินค้า id = 99 }
    Conflict:
      description: ขัดแย้งกับสถานะปัจจุบันของข้อมูล
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: ข้อมูลที่ส่งมาไม่ผ่านการตรวจสอบ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: ต้องระบุ cost_price
              details: { field: cost_price }
  schemas:
...
    StockAdjustRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: integer, minimum: 1 }
        quantity:
          type: integer
          description: ค่าบวก = รับเข้า, ค่าลบ = ตัดออก, ห้ามเป็น 0
          example: -5
        reason: { type: string, maxLength: 255, example: เบิกใช้ภายใน }
        created_by: { type: string, maxLength: 100, example: somchai }
    StockTransaction:
      type: object
      properties:
        id: { type: integer }
        product_id: { type: integer }
        type: { type: string, enum: [IN, OUT] }
        quantity: { type: integer, description: จำนวนบวกเสมอ ทิศทางดูจาก type }
        change: { type: integer, description: จำนวนพร้อมเครื่องหมาย คำนวณให้เพื่อความสะดวก }
        balance_before: { type: integer }
        balance_after: { type: integer }
        reason: { type: string, nullable: true }
        created_by: { type: string, nullable: true }
        transacted_at: { type: string, format: date-time }
    PageMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        total_pages: { type: integer, example: 3 }
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, nullable: true }